% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aerith-package.R
\docType{package}
\name{Aerith-package}
\alias{Aerith-package}
\alias{Aerith}
\alias{aerith}
\title{Pre and post processing of SIP proteomics data}
\description{
Visualisation of peptide isotopic peaks and SIP peptide spectra match (PSM).
Filtration of high quality PSM. Accurate isotopic abundance calculation of peptide and metabolites.
Visualisation of SIP proteomics results.
}
\details{
Aerith streamlines pre-processing, quality control, and post-processing
steps for stable-isotope proteomics workflows, providing plotting helpers,
scoring utilities, and tidy outputs suited for downstream analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xyz1396/Aerith}
  \item Report bugs at \url{https://github.com/xyz1396/Aerith/issues}
}

}
\author{
\strong{Maintainer}: Yi Xiong \email{yixiong@ou.edu} (\href{https://orcid.org/0009-0005-4591-2366}{ORCID})

}
\keyword{internal}
