% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BYion_peak_calculator_DIY}
\alias{BYion_peak_calculator_DIY}
\title{peak calculator of B Y ione from of one peptide using user defined isotopic distribution}
\usage{
BYion_peak_calculator_DIY(AAstr, Atom, Prob)
}
\arguments{
\item{AAstr}{a CharacterVector}

\item{Atom}{a CharacterVector C13 or N15}

\item{Prob}{a NumericVector for its abundance}
}
\description{
peak calculator of B Y ione from of one peptide using user defined isotopic distribution
}
