% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calBYAtomCountAndBaseMass}
\alias{calBYAtomCountAndBaseMass}
\title{Simple calculator of CHONPS atom count and mass without isotope of B Y ions}
\usage{
calBYAtomCountAndBaseMass(AAstrs)
}
\arguments{
\item{AAstr}{a CharacterVector of peptides}
}
\value{
a list of data.frame of CHONPS atom count and each data.frame is for one peptide
}
\description{
Simple calculator of CHONPS atom count and mass without isotope of B Y ions
}
\examples{
peps <- calBYAtomCountAndBaseMass(c("HK~FL","AD!CH","~ILKMV"))
}
