% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calPepPrecursorMass}
\alias{calPepPrecursorMass}
\title{Estimate a representative precursor isotope mass}
\usage{
calPepPrecursorMass(AAstrs, Atom, Probs)
}
\arguments{
\item{AAstrs}{a CharacterVector of peptides}

\item{Atom}{a Character of "C13", "H2", "O18", "N15", or "S34"}

\item{Probs}{a NumericVector with the same length of AAstr for SIP abundances}
}
\value{
a vector of peptide precursor masses
}
\description{
Estimate a representative precursor isotope mass
}
\details{
Uses the nominal shift of the isotope envelope's most abundant
peak and the mean isotope spacing across all atom sources.
This is an estimate of the modal peak mass.
}
\examples{
masses <- calPepPrecursorMass(c("HKFL", "ADCH"), "C13", c(0.2, 0.3))
}
