% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package initialization}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{character string giving the library directory where the package defining the namespace was found.}

\item{pkgname}{character string giving the name of the package.}
}
\description{
Sets up environment variables and package configuration when Aerith is loaded.
}
\keyword{internal}
