% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnload}
\alias{.onUnload}
\title{Package cleanup}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{character string giving the complete path to the package.}
}
\description{
Cleanup when package is unloaded.
}
\keyword{internal}
