% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractPSMfeatures}
\alias{extractPSMfeatures}
\title{extractPSMfeatures extract featueres of top PSMs from multiple .Spe2Pep.txt files}
\usage{
extractPSMfeatures(Spe2PepFilePath, topN, ftFilepath, ThreadNumber = 3L)
}
\arguments{
\item{Spe2PepFilePath}{a full path with .Spe2Pep.txt files in it}

\item{topN}{store top N PSMs of each scan of one .FT2 file}

\item{ftFilepath}{a full path with .FT1 and .FT2 files in it}

\item{ThreadNumber}{read ThreadNumber of FT file at the same time, it will increase ram usage}
}
\value{
the PSMs in a dataframe in a list
}
\description{
extractPSMfeatures extract featueres of top PSMs from multiple .Spe2Pep.txt files
}
\examples{
psm <- extractPSMfeatures(Spe2PepFilePath, topN, ftFilepath, 3)
}
