% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateCFGs}
\alias{generateCFGs}
\title{generateCFGs}
\usage{
generateCFGs(cfgPath, outPath, element)
}
\arguments{
\item{cfgPath}{a full path of .cfg file}

\item{outPath}{a full path for .cfg file output}

\item{element}{a string of element name, "N" for example}
}
\value{
a bool value if generate succeed or not
}
\description{
generateCFGs
}
