% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateOneCFG}
\alias{generateOneCFG}
\title{generateOneCFG}
\usage{
generateOneCFG(cfgPath, outPath, element, pct, center, width)
}
\arguments{
\item{cfgPath}{a full path of .cfg file}

\item{outPath}{a full path for .cfg file output}

\item{element}{a string of element name, "N" for example}

\item{pct}{a integer of element SIP abundance}

\item{center}{a integer of mass window center}

\item{width}{a integer of mass half window width}
}
\value{
a bool value if generate succeed or not
}
\description{
generateOneCFG
}
\examples{
cfg <- system.file("extdata", "SiprosConfig.cfg", package = "Aerith")
tmp <- tempdir()
tmp <- file.path(tmp, "configs")
generateOneCFG(cfg, tmp, "N", 50, 0, 2)
list.files(tmp, full.names = TRUE)
}
