% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFilterThreshold}
\alias{getFilterThreshold}
\title{getFilterThreshold}
\usage{
getFilterThreshold(workingPath, OverallThreshold)
}
\arguments{
\item{workingPath}{a full path with .sip files in it}

\item{OverallThreshold}{FDR thredhold of peptides}
}
\value{
a dataframe about filter threshold and FDR results
}
\description{
getFilterThreshold
}
