% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFilterThresholdTopPSMs}
\alias{getFilterThresholdTopPSMs}
\title{getFilterThresholdTopPSMs get filter threshold of top PSMs of each scan from multiple .sip file}
\usage{
getFilterThresholdTopPSMs(workingPath, OverallThreshold, topN)
}
\arguments{
\item{workingPath}{a full path with .sip files in it}

\item{OverallThreshold}{FDR thredhold of peptides}

\item{topN}{store top N PSMs of each scan of one .FT file}
}
\value{
a dataframe about filter threshold and FDR results
}
\description{
getFilterThresholdTopPSMs get filter threshold of top PSMs of each scan from multiple .sip file
}
