% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{getMZ}
\alias{getMZ}
\title{add MZ to spectra data.frame}
\usage{
getMZ(spectra, charges = 1:2)
}
\arguments{
\item{spectra}{a dataframe of spectra}

\item{charges}{ion charges}
}
\value{
a dataframe of spectra
}
\description{
add MZ to spectra data.frame
}
\examples{
spectra <- precursor_peak_calculator("KHRIP")
spectra <- getMZ(spectra, 1:2)
}
