% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{getPrecursorSpectra}
\alias{getPrecursorSpectra}
\title{Get AAspectra object of precursor from AA sequence with natural SIP abundance}
\usage{
getPrecursorSpectra(AAstr, charges = c(1, 2))
}
\arguments{
\item{AAstr}{Amino acid string}

\item{charges}{Integer vector of charges. Default is 1:2}
}
\value{
AAspectra object
}
\description{
Get AAspectra object of precursor from AA sequence with natural SIP abundance
}
\examples{
getPrecursorSpectra("KHRIP", 1:2)
}
