% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{getRealScanWithCharge}
\alias{getRealScanWithCharge}
\title{Convert one scan in scans with real charges list of ft file to AAspectra class}
\usage{
getRealScanWithCharge(scanNumber, ft)
}
\arguments{
\item{scanNumber}{ScanNumber of one scan}

\item{ft}{Scans list of ft file}
}
\value{
AAspectra object
}
\description{
Convert one scan in scans with real charges list of ft file to AAspectra class
}
\examples{
a <- readAllScanMS2("demo.FT2")
b <- getRealScanWithCharge(18, a)
plot(b)
}
