% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScanFrequency.R
\name{getRetentionTimeAndPrecursorInfo}
\alias{getRetentionTimeAndPrecursorInfo}
\title{get retention time and precursor mass from scans list of ft file}
\usage{
getRetentionTimeAndPrecursorInfo(ft)
}
\arguments{
\item{ft}{Scans list of ft file}
}
\value{
A data.frame of retention time and precursor mass
}
\description{
get retention time and precursor mass from scans list of ft file
}
\examples{
a <- readAllScanMS2("demo.FT2")
b <- getRetentionTimeAndPrecursorMass(a)
a <- readAllScanMS1("demo.FT1")
b <- getRetentionTimeAndPrecursorMass(a)
}
