% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{getSipPrecursorSpectra}
\alias{getSipPrecursorSpectra}
\title{Get AAspectra object  of precursor from AA sequence
with labeled SIP abundance}
\usage{
getSipPrecursorSpectra(AAstr, Atom = "C13", Prob = 0.0107, charges = 1:2)
}
\arguments{
\item{AAstr}{Amino acide string}

\item{Atom}{"C13" or "N15". Default is "C13"}

\item{Prob}{C13 or N15's abundance. Default is 0.0107}

\item{charges}{Integer vector of charges. Default is 1:2}
}
\value{
AAspectra object
}
\description{
Get AAspectra object  of precursor from AA sequence
with labeled SIP abundance
}
\examples{
getSipPrecursorSpectra("KHRIPCDRK", "C13", 0.05, 1:3)
}
