% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIC.R
\name{getTIC}
\alias{getTIC}
\title{get TIC and retention time from scans list of ft file}
\usage{
getTIC(ft)
}
\arguments{
\item{ft}{Scans list of ft file}
}
\value{
A data.frame of TIC, relative TIC and retention time
}
\description{
TIC: total ion chromatogram.
}
\examples{
demo_file <- system.file("extdata", "demo.FT2", package = "Aerith")
a <- readAllScanMS2(demo_file)
b <- getTIC(a)
}
