% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getUnfilteredPeptides}
\alias{getUnfilteredPeptides}
\title{getUnfilteredPeptides}
\usage{
getUnfilteredPeptides(workingPath)
}
\arguments{
\item{workingPath}{a full path with .sip files in it}
}
\value{
a dataframe of unique peptides and whether it is decoy sequence
}
\description{
getUnfilteredPeptides
}
