% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getUnfilteredPSMs}
\alias{getUnfilteredPSMs}
\title{getUnfilteredPSMs}
\usage{
getUnfilteredPSMs(workingPath)
}
\arguments{
\item{workingPath}{a full path with .sip files in it}
}
\value{
a dataframe of unique PSMs and whether it is decoy sequence
}
\description{
getUnfilteredPSMs
}
