% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{plot,AAspectra-method}
\alias{plot,AAspectra-method}
\title{Draw AAspectra MS plot}
\usage{
\S4method{plot}{AAspectra}(x, linewidth = 0.1)
}
\arguments{
\item{AAspectra}{}
}
\value{
a ggplot2 object
}
\description{
Draw AAspectra MS plot
}
\examples{
a <- getPrecursorSpectra("KHRIP", 2)
plot(a)
}
