% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{plotPrecursorAnnotation}
\alias{plotPrecursorAnnotation}
\title{plot precursor annotation}
\usage{
plotPrecursorAnnotation(
  observedSpect,
  pep,
  charge,
  Atom,
  Prob,
  isoCenter = 0,
  isoWidth = 0,
  xwidth = 0,
  ifRemoveNotFoundIon = FALSE,
  ifShowPrecursorChargeAnnotation = TRUE,
  ifShowIsoCenter = TRUE,
  ifShowIsoWindow = TRUE,
  ifAdjustSIPabundance = TRUE,
  breakSize = 1/5,
  linewidth = 0.3
)
}
\arguments{
\item{observedSpect}{AAspectra object of precursor scan}

\item{pep}{peptide sequence}

\item{charge}{precursor charge in consideration}

\item{Atom}{SIP labeled atom "13C" or "15N" for exmaple}

\item{Prob}{its SIP abundance (0.0~1.0)}

\item{isoCenter}{isolation window center. Defaults to the center of
\code{observedSpect} when \code{0}.}

\item{isoWidth}{isolation window width. Defaults to the m/z span of
\code{observedSpect} when \code{0}.}

\item{ifRemoveNotFoundIon}{set it FALSE as default}

\item{ifShowPrecursorChargeAnnotation}{Logical. If TRUE, show precursor charge
annotation labels with ggrepel. Default TRUE.}

\item{ifShowIsoCenter}{Logical. If TRUE, show \code{isoCenter} as a black dashed
vertical reference line. Default TRUE.}

\item{ifShowIsoWindow}{Logical. If TRUE, show grey dashed vertical lines at
isolation window boundaries (\verb{isoCenter ± isoWidth/2}). Default TRUE.}

\item{ifAdjustSIPabundance}{Logical. If TRUE, estimate and re-fit SIP
abundance from initial precursor annotation. Default TRUE.}

\item{breakSize}{Numeric multiplier for isolation window to compute x-axis
break step. Default is 1/5.}

\item{linewidth}{Numeric width of peaks. Default 0.3.}
}
\value{
ggplot2 layer
}
\description{
plot precursor annotation
}
\examples{
realMZ <- c(
  894.9413, 895.4429, 895.9444, 896.3896, 896.4448, 896.9463,
  897.3890, 897.8896, 898.3930, 898.4734, 901.8851, 902.4465,
  902.9483, 903.4498, 903.9504, 910.8968, 911.4449, 912.3784
)
realIntensity <- c(
  16660537.0, 12344664.0, 6128400.5, 1448961.1, 1614148.1, 713238.8,
  1999402.4, 1124157.4, 567865.2, 647140.8, 709644.2, 7805729.0,
  8421993.0, 3200114.2, 1286055.5, 620246.8, 540861.6, 1079918.5
)
realCharge <- c(2, 2, 2, 0, 2, 2, 2, 2, 2, 0, 0, 2, 2, 2, 2, 0, 0, 2)
scan <- list(
  peaks = data.frame(
    mz = realMZ,
    intensity = realIntensity,
    charge = realCharge
  ),
  precursorCharges = 2
)
observedSP <- getRealScanFromList(scan)
p <- plotPrecursorAnnotation(
  observedSpect = observedSP,
  pep = "GITINTSHVEYDTPTR", charge = 2,
  Atom = "C13", Prob = 0.01,
  isoCenter = 902.4471, isoWidth = 5.0, xwidth = 20.0,
  ifRemoveNotFoundIon = TRUE
)
p
}
