% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProteinSipPct.R
\name{plotProSipPct}
\alias{plotProSipPct}
\title{plot the distribution of SIP percent of proteins}
\usage{
plotProSipPct(proPath)
}
\arguments{
\item{proPath}{a pro.cluster.txt file's path}
}
\value{
a ggplot2 obj
}
\description{
plot the distribution of SIP percent of proteins
}
\examples{
\dontrun{
p <- plotProSipPct("demo.pro.cluster.txt")
p
}
}
