% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPSMs.R
\name{plotPSMs}
\alias{plotPSMs}
\title{plot PSMs from FT2 files and PSM results}
\usage{
plotPSMs(
  realScans,
  charges,
  Atom = "C13",
  Probs,
  BYcharge = 1:2,
  ftFileNames,
  scanNumbers,
  pepSeqs,
  proNames,
  path = ""
)
}
\arguments{
\item{realScans}{List of AAspectra objects of real scan}

\item{charges}{Integer vector of precursor charge of PSMs}

\item{Atom}{"C13" or "N15". Default is "C13"}

\item{Probs}{C13 or N15's abundance of PSMs}

\item{BYcharge}{Integer vector of B Y ion. Default is 1:2}

\item{ftFileNames}{Character vector of FT2 file names}

\item{scanNumbers}{Integer vector of scan number of PSMs}

\item{pepSeqs}{Character vector of peptide sequence of PSMs}

\item{proNames}{Character vector of protein name of PSMs}

\item{path}{Output path of pdf. Defaut is "."}
}
\description{
plot PSMs from FT2 files and PSM results
}
\examples{
plotPSMs()
}
