% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoreDistribution.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{plot the score of PSMs at different charge and mass error}
\usage{
plotScoreDistribution(sipFile)
}
\arguments{
\item{sipFile}{a .sip file's path}
}
\value{
a ggplot2 obj
}
\description{
plot the score of PSMs at different charge and mass error
}
\examples{
sipFile <- system.file("extdata", "demo.sip", package = "Aerith")
plotScoreDistribution(sipFile)
}
