% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{plotSipBYionLabel}
\alias{plotSipBYionLabel}
\title{Draw AAspectra MS plot with B Y ion Labels}
\usage{
plotSipBYionLabel(spect)
}
\arguments{
\item{spect}{AAspectra object of B Y ions}
}
\value{
ggplot2 layer
}
\description{
Draw AAspectra MS plot with B Y ion Labels
}
\examples{
a <- getSipBYionSpectra("KHRIPCDRK", "C13", 0.05, 1:2)
p <- plot(a)
p + plotSipBYionLabel(a)
}
