% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIC.R
\name{plotTIC}
\alias{plotTIC}
\title{Plot TIC of MS1 or MS2}
\usage{
plotTIC(tic)
}
\arguments{
\item{tic}{A data.frame of TIC and retention time}
}
\value{
a ggplot2 object
}
\description{
Plot TIC of MS1 or MS2
}
\examples{
a <- readAllScanMS2("demo.FT2")
b <- getTIC(a)
plotTIC(b)
}
