% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precursor_peak_calculator_DIY}
\alias{precursor_peak_calculator_DIY}
\title{Simple peak calculator of user defined isotopic distribution of one peptide}
\usage{
precursor_peak_calculator_DIY(AAstr, Atom, Prob)
}
\arguments{
\item{AAstr}{a CharacterVector}

\item{Atom}{a CharacterVector C13 or N15}

\item{Prob}{a NumericVector for its abundance}
}
\description{
Simple peak calculator of user defined isotopic distribution of one peptide
}
