% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precursor_peak_calculator_DIY_averagine}
\alias{precursor_peak_calculator_DIY_averagine}
\title{Simple peak calculator of user defined isotopic distribution of one peptide by averagine}
\usage{
precursor_peak_calculator_DIY_averagine(AAstrs, Atom, Prob)
}
\arguments{
\item{Atom}{a CharacterVector C13 or N15}

\item{Prob}{a NumericVector for its abundance}

\item{AAstr}{a CharacterVector of peptides}
}
\value{
a list of DataFrames of spectra
}
\description{
Simple peak calculator of user defined isotopic distribution of one peptide by averagine
}
