% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rankyfify}
\alias{rankyfify}
\title{rankify numeric vector via ftFileWriter}
\usage{
rankyfify(a)
}
\arguments{
\item{a}{A numeric vector whose values will be rank-transformed.}
}
\value{
A numeric vector containing the ranks of the input values.
}
\description{
rankify numeric vector via ftFileWriter
}
\examples{
demo_vec <- c(12.5, 3.2, 7.7, 3.2)
rankyfify(demo_vec)
}
