% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readAllScanMS1}
\alias{readAllScanMS1}
\title{read MS1 scans with scanNumber as index}
\usage{
readAllScanMS1(ftFile)
}
\arguments{
\item{ftFile}{a ft1 file's full path}
}
\value{
a list of MS1 scans with names of scan number
}
\description{
read MS1 scans with scanNumber as index
}
\examples{
demo_file <- system.file("extdata", "demo.FT1", package = "Aerith")
ft1 <- readAllScanMS1(demo_file)
}
