% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readAllScanMS2}
\alias{readAllScanMS2}
\title{read MS2 scans with scanNumber as index}
\usage{
readAllScanMS2(ftFile)
}
\arguments{
\item{ftFile}{a ft2 file's full path}
}
\value{
a list of MS2 scans with names of scan number
}
\description{
read MS2 scans with scanNumber as index
}
\examples{
ft2 <- readAllScanMS2("demo.ft2")
}
