% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readFilesScansTopPSMs}
\alias{readFilesScansTopPSMs}
\title{readFilesScansTopPSMs read each scan's top PSMs from multiple .sip files}
\usage{
readFilesScansTopPSMs(workingPath, topN)
}
\arguments{
\item{workingPath}{a full path with .sip files in it}

\item{topN}{store top N PSMs of each scan of one .FT file}
}
\description{
readFilesScansTopPSMs read each scan's top PSMs from multiple .sip files
}
