% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readFTheader}
\alias{readFTheader}
\title{read FT file header}
\usage{
readFTheader(ftFile)
}
\arguments{
\item{ftFile}{a ft1 file's full path}
}
\value{
a list of ft file header
}
\description{
read FT file header
}
\examples{
header <- readFTheader("demo.ft1")
}
