% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMzml.R
\name{readMzmlMS2}
\alias{readMzmlMS2}
\title{Read MS2 spectra from .mzML file}
\usage{
readMzmlMS2(ms)
}
\arguments{
\item{ms}{A .mzML files's path}
}
\value{
A list of MS2 scans with names of scan number
}
\description{
Read MS2 spectra from .mzML file
}
\examples{
# mzR can be installed from bioconductor
# library(mzR)
demo_file <- system.file("extdata", "demo.mzML", package = "Aerith")
a <- readMzmlMS2(demo_file)
}
