% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readOneScanMS1}
\alias{readOneScanMS1}
\title{readOneScanMS1}
\usage{
readOneScanMS1(ftFile, scanCount)
}
\arguments{
\item{ftFile}{a ft1 file's full path}

\item{scanCount}{the scanCount'th scan}
}
\description{
readOneScanMS1
}
