% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readOneScanMS2}
\alias{readOneScanMS2}
\title{readOneScanMS2}
\usage{
readOneScanMS2(ftFile, scanNumber)
}
\arguments{
\item{ftFile}{a ft2 file's full path}

\item{scanNumber}{the scan at scanNumber}
}
\value{
a list of MS2 scan
}
\description{
readOneScanMS2
}
\examples{
ft2 <- readOneScanMS2("demo.FT2", 2)
}
