% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMzml.R
\name{readPepXMLtable}
\alias{readPepXMLtable}
\title{Read PSM table from .pepXML file}
\usage{
readPepXMLtable(pepXML)
}
\arguments{
\item{pepXML}{A .pepXML files's path}
}
\value{
A dataframe of psm table
}
\description{
Read PSM table from .pepXML file
}
\examples{
# mzR can be installed from bioconductor
# library(mzR)
demo_file <- system.file("extdata", "demo.pepXML", package = "Aerith")
a <- readPepXMLtable(demo_file)
}
