% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMzml.R
\name{readPSMtsv}
\alias{readPSMtsv}
\title{Read PSM TSV File}
\usage{
readPSMtsv(tsv)
}
\arguments{
\item{file_path}{A character string specifying the path to the PSM TSV file.}
}
\value{
A data frame containing the data from the PSM TSV file.
}
\description{
This function reads a Peptide-Spectrum Match (PSM) file in TSV (Tab-Separated Values) format.
}
\examples{
demo_file <- system.file("extdata", "demo.psm.txt", package = "Aerith")
a <- readPSMtsv(demo_file)
}
