% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readScansMS2}
\alias{readScansMS2}
\title{read MS2 scans with scanNumber as index in a range}
\usage{
readScansMS2(ftFile, startScanNumber, endScanNumber)
}
\arguments{
\item{ftFile}{a ft2 file's full path}

\item{startScanNumber}{read scans starting from this scanNumber}

\item{endScanNumber}{read scans ending at this scanNumber}
}
\value{
a list of MS2 scans with names of scan number
}
\description{
read MS2 scans with scanNumber as index in a range
}
\examples{
demo_file <- system.file("extdata", "demo.FT2", package = "Aerith")
ft2 <- readScansMS2(demo_file, 1350, 1355)
}
