% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readScansMS2Vector}
\alias{readScansMS2Vector}
\title{read MS2 scans with scanNumber as index in a vector}
\usage{
readScansMS2Vector(ftFile, scanNumbersVector)
}
\arguments{
\item{ftFile}{a ft2 file's full path}

\item{scanNumbersVector}{read scans starting of these scanNumbers}
}
\value{
a list of MS2 scans with names of scan number
}
\description{
read MS2 scans with scanNumber as index in a vector
}
\examples{
demo_file <- system.file("extdata", "demo.FT2", package = "Aerith")
ft2 <- readScansMS2Vector(demo_file, c(1350, 1355, 1359))
}
