% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readSpe2PepFilesScansTopPSMsFromOneFT2}
\alias{readSpe2PepFilesScansTopPSMsFromOneFT2}
\title{readSpe2PepFilesScansTopPSMsFromOneFT2 read each scan's top PSMs from multiple .Spe2PepFile.txt files of one .FT2 file}
\usage{
readSpe2PepFilesScansTopPSMsFromOneFT2(workingPath, pattern, topN = 5L)
}
\arguments{
\item{workingPath}{a full path with .Spe2PepFile.txt files in it}

\item{pattern}{a regex pattern of the .FT2 file}

\item{topN}{store top N PSMs of each scan of one .FT2 file}
}
\value{
a dataframe of top N PSMs
}
\description{
readSpe2PepFilesScansTopPSMsFromOneFT2 read each scan's top PSMs from multiple .Spe2PepFile.txt files of one .FT2 file
}
\examples{
top3 <-  readSpe2PepFilesScansTopPSMsFromOneFT2("testDir", ".*demo1.*", 3)
}
