% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readSpe2Peps}
\alias{readSpe2Peps}
\title{readSpe2Peps}
\usage{
readSpe2Peps(workingPath)
}
\arguments{
\item{workingPath}{a full path with .Spe2Pep.txt files in it}
}
\value{
the PSMs dataframes in lists
}
\description{
readSpe2Peps
}
\examples{
psm <- readSpe2Peps("testDir")
psm <- psm[[1]]$PSM
}
