% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{residue_peak_calculator_DIY}
\alias{residue_peak_calculator_DIY}
\title{Simple residue peak calculator of user defined isotopic distribution of one residue}
\usage{
residue_peak_calculator_DIY(residue, Atom, Prob)
}
\arguments{
\item{residue}{residue name}

\item{Atom}{isotopes of "C13", "N15", "H2", "O18", "S34"}

\item{Prob}{its SIP abundance (0.0~1.0)}
}
\value{
A DataFrame with two columns: "Mass" containing the mass of each isotope and "Prob" containing the probability of each isotope.
}
\description{
Simple residue peak calculator of user defined isotopic distribution of one residue
}
\examples{
df <- residue_peak_calculator_DIY("A", "C13", 0.2)
}
