% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreIntensity}
\alias{scoreIntensity}
\title{scoreIntensity}
\usage{
scoreIntensity(observed, realIntensity, expectedIntensity, Atom, Prob)
}
\arguments{
\item{observed}{this peak is observed or not}

\item{realIntensity}{real intensity in MS2 scan}

\item{expectedIntensity}{expected intensity}

\item{Atom}{"C13" or "N15"}

\item{Prob}{its SIP abundance (0.0~1.0)}
}
\value{
a score of this intensity match
}
\description{
scoreIntensity
}
\examples{
scoreIntensity(TRUE, 1200.0, 1180.0, "C13", 0.02)
}
