% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreIntensityByCE}
\alias{scoreIntensityByCE}
\title{scoreIntensityByCrossEntropy}
\usage{
scoreIntensityByCE(expectedIntensity, observedIntensity)
}
\arguments{
\item{expectedIntensity}{expected intensityreal}

\item{observedIntensity}{observed intensity in MS2 scan}
}
\description{
scoreIntensityByCrossEntropy
}
