% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scorePSMold}
\alias{scorePSMold}
\title{scorePSMold old function of scoreWeightSumHighMS2}
\usage{
scorePSMold(realMZ, realIntensity, realCharge, pepSeq, Atom, Prob)
}
\arguments{
\item{realMZ}{mz vector in MS2 scan}

\item{realIntensity}{intensity vector in MS2 scan}

\item{pepSeq}{a string of peptide}

\item{Atom}{"C13" or "N15"}

\item{Prob}{its SIP abundance (0.0~1.0)}
}
\value{
a score of this PSM
}
\description{
scorePSMold old function of scoreWeightSumHighMS2
}
