% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{show,AAspectra-method}
\alias{show,AAspectra-method}
\title{Show an annotated mass spectrum}
\usage{
\S4method{show}{AAspectra}(object)
}
\arguments{
\item{object}{An \linkS4class{AAspectra} object.}
}
\value{
Returns \code{NULL} invisibly, after printing a compact spectrum summary.
}
\description{
Show an annotated mass spectrum
}
\examples{
show(getPrecursorSpectra("KHRIP", 1:2))
}
