% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeAllScanMS1}
\alias{writeAllScanMS1}
\title{write all MS1 scans has charge}
\usage{
writeAllScanMS1(header, scansList, ftFile)
}
\arguments{
\item{header}{a list of FT file header}

\item{ftFile}{a ft1 file's output path}

\item{scans}{a list of scans for output}
}
\value{
void
}
\description{
write all MS1 scans has charge
}
\examples{
header <- readFTheader("demo.ft1")
ft1 <- readAllScanMS1("demo.ft1")
writeAllScanMS1(header,ft1[1:10],"demo10.ft1")
}
