% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\docType{class}
\name{AAspectra-class}
\alias{AAspectra-class}
\title{AAspectra S4 object}
\description{
AAspectra S4 object
}
\section{Slots}{

\describe{
\item{\code{spectra}}{data.frame.}

\item{\code{charges}}{numericVector.}

\item{\code{AAstr}}{character.}
}}

\examples{
AAstr <- "KHRIP"
spectra <- getPrecursorSpectra(AAstr,1:2)
class(spectra)
}
