% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\docType{class}
\name{AAspectra-class}
\alias{AAspectra-class}
\title{AAspectra S4 object}
\description{
This class is unified data structure for spectra plotting.
The data.frame of spectra consists of columns of "Mass", "Prob"
"Kind", "Charge", and "MZ".
}
\section{Slots}{

\describe{
\item{\code{spectra}}{data.frame for ion peaks}

\item{\code{charges}}{numeric vector for precursor charge}

\item{\code{AAstr}}{characters for amino acid sequence or compound name}
}}

\examples{
AAstr <- "KHRIP"
spectra <- getPrecursorSpectra(AAstr, 1:2)
class(spectra)
}
