% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aerith-package.R
\name{Aerith-package}
\alias{Aerith-package}
\alias{Aerith}
\alias{aerith}
\title{Pre and post processing of SIP proteomics data}
\description{
Quality control of mass spectra raw file. Post-processing of stable isotope probing (SIP) proteomics data from Sipros output.
}
