% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aerith-package.R
\name{Aerith-package}
\alias{Aerith-package}
\alias{Aerith}
\alias{aerith}
\title{Pre and post processing of SIP proteomics data}
\description{
Visualisation of peptide isotopic peaks and SIP peptide spectra match (PSM).
Filtration of high quality PSM. Accurate isotopic abundance calculation of peptide and metabolites.
Visualisation of SIP proteomics results.
}
