% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calPepAtomCount}
\alias{calPepAtomCount}
\title{Simple calculator of CHONPS atom count of peptide}
\usage{
calPepAtomCount(AAstrs)
}
\arguments{
\item{AAstr}{a CharacterVector of peptides}
}
\value{
a dataframe of CHONPS atom count each row is for one peptide
}
\description{
Simple calculator of CHONPS atom count of peptide
}
\examples{
df <- calPepAtomCount(c("HKFL","ADCH"))
}
