% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calPepPrecursorMass}
\alias{calPepPrecursorMass}
\title{Simple calculator of peptide precursor mass by binomial NP}
\usage{
calPepPrecursorMass(AAstrs, Atom, Probs)
}
\arguments{
\item{Atom}{a Character of "C13", "H2", "O18", "N15", or "S34"}

\item{Probs}{a NumericVector with the same length of AAstr for SIP abundances}

\item{AAstr}{a CharacterVector of peptides}
}
\value{
a vector of peptide precursor masses
}
\description{
Simple calculator of peptide precursor mass by binomial NP
}
\examples{
masses <- calPepPrecursorMass(c("HKFL", "ADCH"), "C13", c(0.2, 0.3))
}
