% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{getRealScan}
\alias{getRealScan}
\title{Convert one scan with charges=1 normalized by highest peak in scans list of ft file to AAspectra class}
\usage{
getRealScan(scanNumber, ft)
}
\arguments{
\item{scanNumber}{ScanNumber of one scan}

\item{ft}{Scans list of ft file}
}
\value{
AAspectra object
}
\description{
Convert one scan with charges=1 normalized by highest peak in scans list of ft file to AAspectra class
}
\examples{
demo_file <- system.file("extdata", "demo.FT2", package = "Aerith")
a <- readAllScanMS2(demo_file)
b <- getRealScan(1388, a)
plot(b)
}
