% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getUnfilteredPSMs}
\alias{getUnfilteredPSMs}
\title{getUnfilteredPSMs}
\usage{
getUnfilteredPSMs(sipPath, ftPath, topN)
}
\arguments{
\item{sipPath}{a full path with .sip files in it}

\item{ftPath}{a full path with .ft files in it}

\item{topN}{store top N PSMs of each scan of one .FT file}
}
\description{
getUnfilteredPSMs
}
\examples{
demo_dir <- system.file("extdata", package = "Aerith")
head(getUnfilteredPSMs(demo_dir, demo_dir, 10))
}
