% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{plot.AAspectra}
\alias{plot.AAspectra}
\title{Draw AAspectra MS plot}
\usage{
plot.AAspectra(x, linewidth = 0.1)
}
\arguments{
\item{x}{AAspectra object}

\item{linewidth}{}
}
\value{
a ggplot2 object
}
\description{
Draw AAspectra MS plot
}
\examples{
a <- getPrecursorSpectra("KHRIP", 2)
plot(a)
}
