% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProteinSipPct.R
\name{plotFilteredPCTIntensitySummary}
\alias{plotFilteredPCTIntensitySummary}
\title{Plot decoy-filtered PSMs' PCT and intensity summary by each input file}
\usage{
plotFilteredPCTIntensitySummary(
  psms_dir = "psms/",
  output_file = "decoy_filtered_PCT_and_intensity_summary_by_file.pdf",
  width = 16,
  height = 12
)
}
\arguments{
\item{psms_dir}{Directory containing filtered PSM files (default: "psms/").}

\item{output_file}{Output PDF file name for the plot.}

\item{width}{Width of the output PDF (default: 16).}

\item{height}{Height of the output PDF (default: 12).}
}
\value{
A ggplot2 object representing the hexbin plot.
}
\description{
This function reads all filtered PSM files in from output directory of Sipros 5, combines them, filters for labeled PSMs,
extracts file names, computes log2 intensities, and generates a hexbin plot faceted by each input file.
}
