% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScanFrequency.R
\name{plotPrecursorMzFrequency}
\alias{plotPrecursorMzFrequency}
\title{Plot precursor MZ frequency per 5 per minute of MS2}
\usage{
plotPrecursorMzFrequency(info)
}
\arguments{
\item{info}{A data.frame of retention time and precursor mass}
}
\value{
A ggplot layer of scan frequency of MS2
}
\description{
Plot precursor MZ frequency per 5 per minute of MS2
}
\examples{
a <- readAllScanMS2("demo.FT2")
a <- getRetentionTimeAndPrecursorMass(a)
plotPrecursorMzFrequency(a)
}
