% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAspectra.R
\name{plotPSMannotation}
\alias{plotPSMannotation}
\title{plot PSM annotation}
\usage{
plotPSMannotation(
  observedSpect,
  pep,
  Atom,
  Prob,
  charges,
  isoCenter = 0,
  isoWidth = 0,
  ifRemoveNotFoundIon = False
)
}
\arguments{
\item{pep}{peptide sequence}

\item{Atom}{SIP labeled atom "13C" or "15N" for exmaple}

\item{Prob}{its SIP abundance (0.0~1.0)}

\item{charges}{charges numeric vector for B/Y ions in consideration}

\item{isoCenter}{isolation window center, set it 0 as default if not remove peaks in isolation window}

\item{isoWidth}{isolation window width, set it 0 as default if not remove peaks in isolation window}

\item{ifRemoveNotFoundIon}{set it False as default}

\item{obaservedSpect}{AAspectra object of real scan}
}
\value{
ggplot2 layer
}
\description{
plot PSM annotation
}
\examples{
a <- readAllScanMS2("107728.ft2.FT2")
a <- getRealScan(a, "107728")
p <- plotPSMannotation(
  observedSpect = a,
  pep = "HSQVFSTAEDNQSAVTIHVLQGER", Atom = "C13", Prob = 0.01,
  charges = c(2), isoCenter = 886.65, isoWidth = 4.0,
  ifRemoveNotFoundIon = False
)
p
}
