% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScanFrequency.R
\name{plotScanFrequencyMS2}
\alias{plotScanFrequencyMS2}
\title{Plot scan frequency layer of MS2}
\usage{
plotScanFrequencyMS2(info)
}
\arguments{
\item{info}{A data.frame of retention time and precursor mass}
}
\value{
A ggplot layer of scan frequency of MS2
}
\description{
Plot scan frequency layer of MS2
}
\examples{
a <- readAllScanMS2("demo.FT2")
a <- getRetentionTimeAndPrecursorMass(a)
b <- readAllScanMS1("demo.FT1")
b <- getRetentionTimeAndPrecursorMass(b)
plotScanFrequency(a) + plotScanFrequencyMS2(b)
}
