% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precursor_peak_calculator_DIY}
\alias{precursor_peak_calculator_DIY}
\title{Precursor Peak Calculator with User-Defined Isotopic Distribution}
\usage{
precursor_peak_calculator_DIY(AAstr, Atom, Prob)
}
\arguments{
\item{AAstr}{A string representing the amino acid sequence.}

\item{Atom}{A string representing the isotope ("C13", "N15", "H2", "O18", "S34").}

\item{Prob}{A double representing the abundance of the specified isotope (0.0 to 1.0).}
}
\value{
A DataFrame with two columns: "Mass" containing the mass of each isotope and "Prob" containing the probability of each isotope.
}
\description{
This function calculates the isotopic distribution of a given amino acid string with a user-defined isotopic distribution and returns a DataFrame containing the mass and probability of each isotope.
}
\examples{
# Example usage
df <- precursor_peak_calculator_DIY("PEPTIDE", "C13", 0.2)
df <- precursor_peak_calculator_DIY("PEPTIDE", "N15", 0.5)
}
