% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precursor_peak_calculator_DIY_averagine}
\alias{precursor_peak_calculator_DIY_averagine}
\title{Simple peak calculator of user defined isotopic distribution of one peptide by averagine}
\usage{
precursor_peak_calculator_DIY_averagine(AAstrs, Atom, Prob)
}
\arguments{
\item{AAstrs}{a CharacterVector of peptides}

\item{Atom}{a CharacterVector C13 or N15}

\item{Prob}{a NumericVector for its abundance}
}
\value{
a list of DataFrames of spectra
}
\description{
Simple peak calculator of user defined isotopic distribution of one peptide by averagine
}
\examples{
demoSpectra <- precursor_peak_calculator_DIY_averagine(c("PEPTIDE", "ACDEFGHIK"), "C13", 0.25)
demoSpectra[[1]]
}
