% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMzml.R
\name{readMzmlMS1}
\alias{readMzmlMS1}
\title{Read MS1 spectra from .mzML file}
\usage{
readMzmlMS1(ms)
}
\arguments{
\item{ms}{A .mzML files's path}
}
\value{
A list of MS1 scans with names of scan number
}
\description{
Read MS1 spectra from .mzML file
}
\examples{
# mzR can be installed from bioconductor
library(mzR)
demo_file <- system.file("extdata", "demo.mzML", package = "Aerith")
a <- readMzmlMS1(demo_file)
}
