% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readOneScanMS1}
\alias{readOneScanMS1}
\title{readOneScanMS1}
\usage{
readOneScanMS1(ftFile, scanNumber)
}
\arguments{
\item{ftFile}{a ft1 file's full path}

\item{scanNumber}{the scan at scanNumber}
}
\value{
a list of MS1 scan
}
\description{
readOneScanMS1
}
\examples{
ft1 <- readOneScanMS1("demo.FT1", 2)
}
