% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readOneScanMS2}
\alias{readOneScanMS2}
\title{readOneScanMS2}
\usage{
readOneScanMS2(ftFile, scanCount)
}
\arguments{
\item{ftFile}{a ft2 file's full path}

\item{scanCount}{the scanCount'th scan}
}
\description{
readOneScanMS2
}
