% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readScansMS2}
\alias{readScansMS2}
\title{readScansMS2}
\usage{
readScansMS2(ftFile, scanCount)
}
\arguments{
\item{ftFile}{a ft2 file's full path}

\item{scanNumber}{the scanNumber th scan}
}
\description{
readScansMS2
}
