% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readSpe2Pep}
\alias{readSpe2Pep}
\title{readSpe2Pep}
\usage{
readSpe2Pep(Spe2PepFile)
}
\arguments{
\item{Spe2PepFile}{a .Spe2PepFile file's full path}
}
\value{
the PSMs in a dataframe in a list
}
\description{
readSpe2Pep
}
\examples{
target_file <- system.file("extdata", "demo_target.Spe2Pep.txt", package = "Aerith")
psm <- readSpe2Pep(target_file)
psm <- psm$PSM
}
