% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readSpe2PepFilesScansTopPSMs}
\alias{readSpe2PepFilesScansTopPSMs}
\title{readSpe2PepFilesScansTopPSMs read each scan's top PSMs from multiple .Spe2Pep.txt files}
\usage{
readSpe2PepFilesScansTopPSMs(workingPath, topN = 5L)
}
\arguments{
\item{workingPath}{a full path with .Spe2Pep.txt files in it}

\item{topN}{store top N PSMs of each scan of one .FT2 file}
}
\value{
the PSMs in a dataframe in a list
}
\description{
readSpe2PepFilesScansTopPSMs read each scan's top PSMs from multiple .Spe2Pep.txt files
}
\examples{
psm <- readSpe2PepFilesScansTopPSMs("testDir")
}
