% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readSpe2Peps}
\alias{readSpe2Peps}
\title{readSpe2Peps}
\usage{
readSpe2Peps(workingPath)
}
\arguments{
\item{workingPath}{a full path with .Spe2Pep.txt files in it}
}
\value{
the PSMs dataframes in lists
}
\description{
readSpe2Peps
}
\examples{
tmp <- tempdir()
sip_dir <- file.path(tmp, "sip")
dir.create(sip_dir)
demo_file <- system.file("extdata", "demo_target.Spe2Pep.txt", package = "Aerith")
file.copy(demo_file, file.path(sip_dir, "Pan_052322_X13.SIP_C13_050_000target.Spe2Pep.txt"))
demo_file <- system.file("extdata", "demo_decoy.Spe2Pep.txt", package = "Aerith")
file.copy(demo_file, file.path(sip_dir, "Pan_052322_X13.SIP_C13_050_000decoy.Spe2Pep.txt"))
list.files(sip_dir, full.names = TRUE)
psm <- readSpe2Peps(sip_dir)
psm <- psm[[1]]$PSM
}
