% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{residue_peak_calculator_DIY}
\alias{residue_peak_calculator_DIY}
\title{Simple residue peak calculator of user defined isotopic distribution of one residue}
\usage{
residue_peak_calculator_DIY(residue, Atom, Prob)
}
\arguments{
\item{residue}{residue name}

\item{Atom}{"C13" or "N15"}

\item{Prob}{its SIP abundance (0.0~1.0)}
}
\description{
Simple residue peak calculator of user defined isotopic distribution of one residue
}
