% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeAllScanMS2}
\alias{writeAllScanMS2}
\title{write all MS2 scans has charge}
\usage{
writeAllScanMS2(header, scansList, ftFile)
}
\arguments{
\item{header}{a list of FT file header}

\item{ftFile}{a ft2 file's output path}

\item{scans}{a list of scans for output}
}
\value{
void
}
\description{
write all MS2 scans has charge
}
\examples{
header <- readFTheader("demo.ft2")
ft2 <- readAllScanMS2("demo.ft2")
writeAllScanMS2(header,ft2[1:10],"demo10.ft2")
}
